/hubblenetwork/cli.py: Added ble detect function
Added a new cli function to find packets at a given key, then dumps a
JSON of packet data if successful, or an error if not. There is also
a debug mode that displays logs in the console for development. The
current intention is to use this in Actions for v1 of the endpoint
HITL setup

Signed off by Hunter Patchett: hunter@hubble.com